+++ hydra-cardano-api/src/Hydra/Cardano/Api/PlutusScript.hs
@@ -5,6 +5,7 @@ module Hydra.Cardano.Api.PlutusScript where
 import Hydra.Cardano.Api.Prelude
 
 import Cardano.Ledger.Alonzo.Scripts qualified as Ledger
+import Cardano.Ledger.Babbage.Scripts qualified as Ledger
 import Cardano.Ledger.Plutus.Language qualified as Ledger
 import Data.ByteString.Short qualified as SBS
 import PlutusLedgerApi.Common qualified as Plutus
@@ -18,10 +19,15 @@ import Test.QuickCheck (listOf)
 --
 -- (a) If the given script is a timelock script, it throws an impure exception;
 -- (b) If the given script is in a wrong language, it silently coerces it.
-fromLedgerScript :: HasCallStack => Ledger.AlonzoScript era -> PlutusScript lang
+fromLedgerScript ::
+  ( HasCallStack
+  , Ledger.AlonzoEraScript era
+  ) =>
+  Ledger.AlonzoScript era ->
+  PlutusScript lang
 fromLedgerScript = \case
   Ledger.TimelockScript{} -> error "fromLedgerScript: TimelockScript"
-  Ledger.PlutusScript (Ledger.Plutus _ (Ledger.BinaryPlutus bytes)) -> PlutusScriptSerialised bytes
+  Ledger.PlutusScript x -> Ledger.withPlutusScript x (\(Ledger.Plutus (Ledger.PlutusBinary bytes)) -> PlutusScriptSerialised bytes)
 
 -- | Convert a cardano-api 'PlutusScript' into a cardano-ledger 'Script'.
 toLedgerScript ::
@@ -30,11 +36,10 @@ toLedgerScript ::
   PlutusScript lang ->
   Ledger.AlonzoScript (ShelleyLedgerEra Era)
 toLedgerScript (PlutusScriptSerialised bytes) =
-  let lang = case plutusScriptVersion @lang of
-        PlutusScriptV1 -> Ledger.PlutusV1
-        PlutusScriptV2 -> Ledger.PlutusV2
-        PlutusScriptV3 -> Ledger.PlutusV3
-   in Ledger.PlutusScript $ Ledger.Plutus lang (Ledger.BinaryPlutus bytes)
+  Ledger.PlutusScript $ case plutusScriptVersion @lang of
+    PlutusScriptV1 -> Ledger.BabbagePlutusV1 $ Ledger.Plutus (Ledger.PlutusBinary bytes)
+    PlutusScriptV2 -> Ledger.BabbagePlutusV2 $ Ledger.Plutus (Ledger.PlutusBinary bytes)
+    PlutusScriptV3 -> error "toLedgerScript: PlutusV3 not supported in Babbage"
 
 -- | Convert a serialized plutus script into a cardano-api 'PlutusScript'.
 fromPlutusScript :: Plutus.SerialisedScript -> PlutusScript lang