+++ hydra-cardano-api/src/Hydra/Cardano/Api/TxBody.hs
@@ -3,7 +3,8 @@ module Hydra.Cardano.Api.TxBody where
 import Hydra.Cardano.Api.Prelude
 
 import Cardano.Ledger.Alonzo.TxWits qualified as Ledger
-import Cardano.Ledger.Babbage.Tx qualified as Ledger
+import Cardano.Ledger.Api (AlonzoPlutusPurpose (..), AsIndex, AsItem (..), PlutusPurpose)
+import Cardano.Ledger.Babbage.Core (redeemerPointer)
 import Cardano.Ledger.BaseTypes (strictMaybeToMaybe)
 import Cardano.Ledger.Core qualified as Ledger
 import Cardano.Ledger.Plutus.Data qualified as Ledger
@@ -11,29 +12,27 @@ import Data.List (find)
 import Data.Map qualified as Map
 import Hydra.Cardano.Api.PlutusScript (fromLedgerScript)
 import Hydra.Cardano.Api.PolicyId (toLedgerPolicyID, toLedgerScriptHash)
-import Hydra.Cardano.Api.ScriptData (FromScriptData)
 import Hydra.Cardano.Api.TxIn (toLedgerTxIn)
 import PlutusLedgerApi.V2 qualified as Plutus
 
 -- | Find and deserialise from 'ScriptData', a redeemer from the transaction
 -- associated to the given input.
 findRedeemerSpending ::
-  FromScriptData a =>
+  Plutus.FromData a =>
   Tx Era ->
   TxIn ->
   Maybe a
 findRedeemerSpending (getTxBody -> ShelleyTxBody _ body _ scriptData _ _) txIn = do
-  ptr <- strictMaybeToMaybe $ Ledger.rdptr body (Ledger.Spending $ toLedgerTxIn txIn)
+  ptr <- strictMaybeToMaybe $ redeemerPointer body (AlonzoSpending . AsItem $ toLedgerTxIn txIn)
   lookupRedeemer ptr scriptData
 
 findRedeemerMinting ::
-  forall a.
-  FromScriptData a =>
+  Plutus.FromData a =>
   Tx Era ->
   PolicyId ->
   Maybe a
 findRedeemerMinting (getTxBody -> ShelleyTxBody _ body _ scriptData _ _) pid = do
-  ptr <- strictMaybeToMaybe $ Ledger.rdptr body (Ledger.Minting $ toLedgerPolicyID pid)
+  ptr <- strictMaybeToMaybe $ redeemerPointer body (AlonzoMinting . AsItem $ toLedgerPolicyID pid)
   lookupRedeemer ptr scriptData
 
 findScriptMinting ::
@@ -53,12 +52,9 @@ findScriptMinting (getTxBody -> ShelleyTxBody _ _ scripts _ _ _) pid = do
 --
 
 lookupRedeemer ::
-  forall a era.
-  ( FromScriptData a
-  , Ledger.Era (ShelleyLedgerEra era)
-  ) =>
-  Ledger.RdmrPtr ->
-  TxBodyScriptData era ->
+  Plutus.FromData a =>
+  PlutusPurpose AsIndex LedgerEra ->
+  TxBodyScriptData Era ->
   Maybe a
 lookupRedeemer ptr scriptData = do
   (d, _exUnits) <- Map.lookup ptr redeemers